+++ macos.c
@@ -140,8 +140,8 @@ static int make_package(int format,            /* I - Format */
         pkgname[1024];   /* Package name */
     file_t *file;        /* Current distribution file */
     command_t *c;        /* Current command */
-    struct passwd *pwd;  /* Pointer to user record */
-    struct group *grp;   /* Pointer to group record */
+    uid_t uid;            /* Resolved owner ID */
+    gid_t gid;            /* Resolved group ID */
     char current[1024];  /* Current directory */
     const char *option;  /* Init script option */
 
@@ -238,11 +238,8 @@ static int make_package(int format,            /* I - Format */
          * Find the username and groupname IDs...
          */
 
-        pwd = getpwnam(file->user);
-        grp = getgrnam(file->group);
-
-        endpwent();
-        endgrent();
+        uid = get_uid(file->user);
+        gid = get_gid(file->group);
 
         /*
          * Copy the file or make the directory or make the symlink as needed...
@@ -256,8 +253,7 @@ static int make_package(int format,            /* I - Format */
             if (Verbosity > 1)
                 printf("%s -> %s...\n", file->src, filename);
 
-            if (copy_file(filename, file->src, file->mode, pwd ? pwd->pw_uid : 0,
-                          grp ? grp->gr_gid : 0))
+            if (copy_file(filename, file->src, file->mode, uid, gid))
                 return (1);
             break;
         case 'i':
@@ -268,8 +264,7 @@ static int make_package(int format,            /* I - Format */
             if (Verbosity > 1)
                 printf("%s -> %s...\n", file->src, filename);
 
-            if (copy_file(filename, file->src, file->mode, pwd ? pwd->pw_uid : 0,
-                          grp ? grp->gr_gid : 0))
+            if (copy_file(filename, file->src, file->mode, uid, gid))
                 return (1);
 
             snprintf(filename, sizeof(filename),
@@ -328,8 +323,7 @@ static int make_package(int format,            /* I - Format */
             if (Verbosity > 1)
                 printf("Directory %s...\n", filename);
 
-            make_directory(filename, file->mode, pwd ? pwd->pw_uid : 0,
-                           grp ? grp->gr_gid : 0);
+            make_directory(filename, file->mode, uid, gid);
             break;
         case 'l':
             staged_path(filename, sizeof(filename), directory, prodfull, file->dst);