+++ setld.c
@@ -48,8 +48,8 @@ make_setld(const char *prodname,     /* I - Product short name */
     file_t *file;          /* Current distribution file */
     command_t *c;          /* Current command */
     char current[1024];    /* Current directory */
-    struct passwd *pwd;    /* Pointer to user record */
-    struct group *grp;     /* Pointer to group record */
+    uid_t uid;              /* Resolved owner ID */
+    gid_t gid;              /* Resolved group ID */
     const char *runlevels; /* Run levels */
 
     REF(platform);
@@ -343,11 +343,8 @@ make_setld(const char *prodname,     /* I - Product short name */
          * Find the username and groupname IDs...
          */
 
-        pwd = getpwnam(file->user);
-        grp = getgrnam(file->group);
-
-        endpwent();
-        endgrent();
+        uid = get_uid(file->user);
+        gid = get_gid(file->group);
 
         /*
          * Copy the file or make the directory or make the symlink as needed...
@@ -362,8 +359,7 @@ make_setld(const char *prodname,     /* I - Product short name */
             if (Verbosity > 1)
                 printf("%s -> %s...\n", file->src, filename);
 
-            if (copy_file(filename, file->src, file->mode, pwd ? pwd->pw_uid : 0,
-                          grp ? grp->gr_gid : 0))
+            if (copy_file(filename, file->src, file->mode, uid, gid))
                 return (1);
             break;
 
@@ -373,8 +369,7 @@ make_setld(const char *prodname,     /* I - Product short name */
             if (Verbosity > 1)
                 printf("Directory %s...\n", filename);
 
-            make_directory(filename, file->mode, pwd ? pwd->pw_uid : 0,
-                           grp ? grp->gr_gid : 0);
+            make_directory(filename, file->mode, uid, gid);
             break;
 
         case 'l':